Re-style the WalletModal panel
... also merge `AccountBalances` with `WalletModal`
